phis_seg_doc/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
blinker==1.6.3
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.4
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
doctr==1.9.0
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.12.4
flask==3.0.0
fonttools==4.43.1
fsspec==2023.9.2
greenlet==3.0.1
h11==0.14.0
h5py==3.9.0
html5lib==1.1
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
langdetect==1.0.9
MarkupSafe==2.1.3
matplotlib==3.8.0
mplcursors==0.5.2
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
postgresql-orm==0.5.8
psycopg2-binary==2.9.9
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
pydyf==0.8.0
pyparsing==3.1.1
pypdfium2==4.20.0
pyphen==0.14.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-doctr==0.7.0
python-multipart==0.0.6
PyYAML==6.0.1
rapidfuzz==3.3.1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
shapely==2.0.1
six==1.16.0
sklearn==0.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
tinycss2==1.2.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing-extensions==4.8.0
Unidecode==1.3.7
urllib3==2.0.6
uvicorn==0.23.2
weasyprint==60.1
webencodings==0.5.1
werkzeug==3.0.0
zipp==3.17.0
zopfli==0.2.3
